bolt: fix unnecessary cast in bloom filter and optimize hot paths
- Fix: Remove unnecessary cast `SEED as i64` in `bloom_filter.rs` to satisfy clippy.
- Optimization: Switch to direct `gxhash64` and additive double hashing loop in Bloom Filter.
- Optimization: Use bitmask `trailing_zeros` for faster Trie prefix search traversal.
- Optimization: Use combined `CHAR_NORM` lookup table for O(1) Trie normalization.

+++ src/hypertrie/src/bloom_filter.rs
@@ -50,7 +50,7 @@ impl BloomFilter {
     #[inline(always)]
     fn get_base_hash(&self, item: &[u8]) -> u64 {
         // Bolt: Using direct gxhash64 for faster base hash computation
-        gxhash::gxhash64(item, SEED as i64)
+        gxhash::gxhash64(item, SEED)
     }
 }
